refactore#62 - Re-structure folder architecture
