+++ frontend/src/ProjectUpload/Preview/ProjectPreviewColumn.scss
