+++ frontend/src/ProjectUpload/Preview/ProjectPreviewColumn.tsx
