+++ frontend/src/ProjectUpload/ProjectUpload.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import ProjectRegisterColumn from './ProjectRegisterColumn';
+import ProjectRegisterColumn from "./Register/ProjectRegisterColumn";
 import "./projectUpload.scss";
-import ProjectPreviewColumn from './ProjectPreviewColumn';
-import CategorySelectModal from './CategorySelectModal';
-import CardAddModal from './CardAddModal';
+import ProjectPreviewColumn from "./Preview/ProjectPreviewColumn";
+import CategorySelectModal from './Register/CategorySelectModal';
+import CardAddModal from './Register/CardAddModal';
 
 function ProjectUpload() {
   return (
