+++ frontend/src/ProjectUpload/Register/ProjectRegisterColumn.scss
