+++ frontend/src/ProjectUpload/Register/projectRegisterColumn.test.tsx
@@ -1,5 +1,11 @@
-import { formatKoreanCurrency } from "../utils/commonFunction";
-import { CARD_ADD_MODAL, CATEGORY_MODAL_OPEN_BTN_VISITED_TEXT, CATEGORY_SELECT_MODAL, KOREA_MONEY_MAX_LENGTH, TOO_BIG_NUMBER_TEXT } from "../utils/commonVariable";
+import { formatKoreanCurrency } from "../../utils/commonFunction";
+import {
+  CARD_ADD_MODAL,
+  CATEGORY_MODAL_OPEN_BTN_VISITED_TEXT,
+  CATEGORY_SELECT_MODAL,
+  KOREA_MONEY_MAX_LENGTH,
+  TOO_BIG_NUMBER_TEXT,
+} from "../../utils/commonVariable";
 import {
   markBtnAsVisited,
   openModal,
